Remove raise statements from Postgres SQL.
These statements are executed regardless of logging level, and however
minor it may be we really don’t want any unnecessary performance hits
in functions that are going be run on many cells of data.

We’re keeping the raised exception in enclave_verify_and_decrypt
because that’s necessary to explain what’s going on, and additionally
that function will only ever be used manually, and is there for
convenience. The statement in our first anonymous function is also kept
for explanatory purposes.