Properly handle encryption/decryption of byeta
We used to coerce all data to the text representation and then run it
through decode, but you can’t do that with bytea due to the way
conversions between bytea and text are handled by Postgres. So we’ve
had to add some special case handling for the type. All other types
should be fine, however, as they’re all guaranteed to have text
representations.